www-server/requirements.txt
gunicorn==19.9.0

django-cors-headers==3.1.1
django-extensions==2.2.8

graphene-django==2.10.1
git+git://github.com/graphql-python/graphene-sqlalchemy.git#master

SQLAlchemy-Utils==0.34.2
psycopg2==2.7.4

spacy==2.1.8

boto3==1.9.226

celery==4.3.0
flower==0.9.3
tornado==4.2
gevent==1.4.0

elasticsearch-dsl==7.1.0
